Fix escaped resource URL name send by some webdav servers.
The problem appears when Neon searches the lock token to enable the
PUT method using the document URL, during the save operation.

Neon compares the path-abempty character by character, so that
path-abempty as:

/alfresco/webdav/Sites/documentLibrary/Lorem%20%27%20ipsum-ab.odt

and

/alfresco/webdav/Sites/documentLibrary/Lorem%20'%20ipsum-ab.odt

come out as different, whereas they are referring to the same resource.

The first version of the path is the one that is formed by using the
URL returned by the WebDAV server when enumerating a collection
(a folder).
This happens when the user navigate to the folder and then selects the
file by double-click.

The second version of the path is the one that is formed by LO while
performing the save.

The added patch addresses it by unescaping and then escaping again the
URL returned while enumerating.

Change-Id: I6af34d206b6a422da85cda39dadd21bb5e51aa58