DEBUG only
Unfortunately it's not possible to detect correctly a
situation where HEAD is not needed to check for a resource.

It would be possible if OPTIONS were used correctly everywhere
but it's not so. OpenKM doesn't work that way: if a file doesn't
exist OPTIONS returns 404, su the resource is interpreted as
'NOT DAV', as it were a Web site, leading to problem when the
file is saved the first time on server.

Change-Id: Ic04db4394e1f720e5f53a5d32df9be7081136099