tdf101094 (??) Use PROPFIIND instead of OPTIONS for a first check
Was trying to overcome the problem with endless redirections, but it
seems that other servers respond to PROFIND in the same way.
Instead of telling 501 or 403 or similar, they redirect in a stupid way
(see picture
<https://wiki.documentfoundation.org/images/3/34/Startcenter-remove-recent-file.png>
PROPFIND gets redirected to:
<https://wiki.documentfoundation.org/Images/3/34/Startcenter-remove-recent-file.png>
A non existing resource.

With OPTIONS there is an endless redirection while opending buggy
document of tdf#99499

Change-Id: I10a72c3b75f3cc8778feb31b91f2c35c75df6244