tdf#101094 (27): Add a mutex guard to OPTIONS neon dispatch calls
This is a sort of follow up to commit:
510da29131e56ba0e1783e505e7131e6f7e0d6b4:

"forced to make certain neon api calls thread-safe due to cups

rhbz#637738 libgcrypt that's been initialized already by cups to be
non-threadsafe is therefore then also non-threadsafe. neon uses
libgcrypt. So in the (as this is the usual circumstance) case that
cups is initialized before neon, then libgcrypt is not threadsafe
even though neon attempt to initialize it as such."

Change-Id: I1f868642ac15bf9fa25d3edd46390467d8fce1f5
Reviewed-on: https://gerrit.libreoffice.org/28401
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Giuseppe Castagno <giuseppe.castagno@acca-esse.eu>
(cherry picked from commit 72589082945e5a197f3fb4b884f48dd4ae275e5f)